requirements.txt
flask==3.0.3
flask-cors==5.0.0
supabase==2.7.0
python-dotenv==1.0.1
openai-whisper==20231117
google-generativeai==0.7.2
Pillow==10.3.0